chore: carousel interface
