+++ packages/taro-components-rn/src/components/Swiper/PropsType.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { ViewStyle } from 'react-native'
+import { StyleProp, ViewStyle } from 'react-native'
 
 export type EventOnChange = {
   detail: {
@@ -26,4 +26,27 @@ export interface SwiperProps {
   vertical?: boolean;
   onChange?: (evt: EventOnChange) => void;
   onAnimationFinish?: (evt: EventOnAnimationFinish) => void;
+}
+
+export interface PaginationProps {
+  vertical?: boolean;
+  current: number;
+  count: number;
+  styles: any;
+  dotStyle?: StyleProp<ViewStyle>;
+  dotActiveStyle?: StyleProp<ViewStyle>;
+}
+
+export interface CarouselProps {
+  selectedIndex?: number;
+  dots?: boolean;
+  vertical?: boolean;
+  autoplay?: boolean;
+  autoplayInterval?: number;
+  infinite?: boolean;
+  style?: StyleProp<ViewStyle>;
+  dotStyle?: StyleProp<ViewStyle>;
+  dotActiveStyle?: StyleProp<ViewStyle>;
+  pagination?: (props: PaginationProps) => React.ReactNode;
+  afterChange?: (index: number) => void;
 }
